Add keys to gutterContainer and scrollContainer to avoid recycling issue
Previously, when the gutter container was removed due to the editor
becoming mini, the lack of keys caused the gutter to be updated with the
recycled cursors vnode. But then we tried to remove the cursors vnode
not realizing it had been moved and tore down all the references.

We probably need to revisit whether it makes sense to recycle vnodes.