WIP: Restore editor scroll position across reloads
It works across one reload, but if you scroll an editor, hide it by
changing tabs, then reload *twice* the scroll position is not preserved.

I think we need to decouple this from the firstVisibleRow concept and
instead have a scrollTopRow / scrollLeftColumn concept. If we have a
pending value when we call getScrollTopColumn/LeftRow, we'll return that
instead of the computed value. We could also allow floats for these
values and use them only for serializing the scroll position or setting
the scroll top / left from the model.